// src/app/client/dataCC.interface.ts
export interface dataCC {
  cardNumberCC: any;
  nomeCC: any;
  sexCC: any;
  heightCC: any;
  nationalityCC: any,
  birthdateCC: any;
  expiricyDateCC: any;
  localOfEmissionCC: any;
  fathersNameCC: any;
  mothersNameCC: any;
  
  nifCC: any;
  socialSecurityCC: any;
  healthNumberCC: any;
  signatureCC: any;
  
  addressCC: any;
  postalCodeCC: any;

  countryCC: any; //Country issuer of the document
}
 